% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\name{outbreaker.package}
\alias{outbreaker.package}
\title{outbreaker: disease outbreak reconstruction using epidemiological and genetic data}
\description{
This package implements the model introduced by Jombart et al. (PLoS Comput. Biol, 2014) for disease outbreak reconstruction using epidemiological and genetic data.
}
\details{
Check tutorials and documentation at:
\url{https://sites.google.com/site/therepiproject/r-pac/outbreaker}
}
\author{
Thibaut Jombart \email{t.jombart@imperial.ac.uk}
}

